requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
networkx>=3.1
weaviate-client>=4.3.0
openai>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-Levenshtein>=0.21.0
fastapi>=0.95.0
uvicorn>=0.22.0
prometheus-client>=0.17.0
dask>=2023.8.0
jsonlines>=3.1.0
joblib>=1.3.0
more-itertools>=9.1.0
click>=8.1.0
loguru>=0.7.0
mmh3>=4.0.0
community>=1.0.0
leidenalg>=0.9.0
pyarrow>=12.0.0
matplotlib-venn>=0.11.9
typing-extensions>=4.7.0
orjson>=3.9.0
hjson>=3.1.0
memory-profiler>=0.61.0
